Fixed naming issues
It turns out that the Condition handler does an explicity
query against \w+ for the name of the Condition module, which
means modules containing :: in their name simply do not work.
On top of that, it doesn't check the regex match for success
so it just blows up if you supply it one.

Also this module has no support for ticket creation, so setting
the Custom Field on startup does not impact the initial value of
the SLA, which means it will always be created with the default
or a manual selection, and may differ from the expected if you
set the CF on creation.

Also got the tests to actually work against the database, but
have not finished writing them.  I ran into an issue where my
database was out of sync with my RT version.